#131 Move Google Maps review data to valueobject
